+++ NEWS.md
@@ -1,3 +1,7 @@
+# ctrdata 1.25.0.9000
+
+New development version
+
 # ctrdata 1.25.0
 
 ## Possibly breaking down-stream workflows